Refactored code organization
