+++ heron/common/tests/python/network/protocol_unittest.py
@@ -15,7 +15,7 @@
 # pylint: disable=missing-docstring
 import unittest2 as unittest
 from heron.common.src.python.network import REQID, HeronProtocol, IncomingPacket
-import heron.common.tests.python.mock_generator as mock_generator
+import heron.common.tests.python.network.mock_generator as mock_generator
 
 class ProtocolTest(unittest.TestCase):
   def setUp(self):
